+++ src/reducers/deviceReducer.ts
@@ -11,6 +11,8 @@ import { DeviceState, RootState } from './types';
 const initialState: DeviceState = {
     board: null,
     isReady: false,
+    serialports: [],
+    selectedSerialport: null,
 };
 
 const deviceSlice = createSlice({
@@ -18,28 +20,46 @@ const deviceSlice = createSlice({
     initialState,
     reducers: {
         deviceDeselected(state) {
+            state.selectedSerialport = null;
+            state.serialports = [];
             state.isReady = initialState.isReady;
         },
-        deviceReady(state) {
-            state.isReady = true;
+        setDeviceReady(state, action) {
+            state.isReady = action.payload;
         },
         dtmBoardSelected(state, action) {
-            state.board = action.payload;
+            state.board = action.payload.board;
+            state.serialports = action.payload.serialports;
+            state.selectedSerialport = action.payload.serialports[0] || null;
+        },
+        serialportSelected(state, action) {
+            state.selectedSerialport = action.payload;
         },
     },
 });
 
 export default deviceSlice.reducer;
 
-const { deviceDeselected, deviceReady, dtmBoardSelected } = deviceSlice.actions;
+const {
+    deviceDeselected,
+    setDeviceReady,
+    dtmBoardSelected,
+    serialportSelected,
+} = deviceSlice.actions;
 
 const getBoard = (state: RootState) => state.app.device.board;
 const getIsReady = (state: RootState) => state.app.device.isReady;
+const getSerialports = (state: RootState) => state.app.device.serialports;
+const getSelectedSerialport = (state: RootState) =>
+    state.app.device.selectedSerialport;
 
 export {
     deviceDeselected,
-    deviceReady,
+    setDeviceReady,
     dtmBoardSelected,
+    serialportSelected,
     getBoard,
     getIsReady,
+    getSerialports,
+    getSelectedSerialport,
 };